+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+- [PR-77](https://github.com/AGH-CEAI/aegis_gym/pull/77) - Added template script for hyperparameter optimization.
+
 ### Changed
 
 - [PR-76](https://github.com/AGH-CEAI/aegis_gym/pull/76) - Increased genesis logging level from "warning" to "info".